Made Anorm the default database engine, and updated documentation appropriately; also made schema installation automatic
