+++ doc/ubuntu-quickstart.txt
@@ -74,11 +74,23 @@ bin/sbt <enter>
 cd <zookeeper install location>
 bin/zkServer.sh start
 
-11. Run it! (note that the default configs don't use zookeeper)
+11. Change the default configs to use Zookeeper and Cassandra
+
+Zipkin runs three daemons: collector, query, and web. The collector and query
+daemons access the database. The default configuration for those two daemons is
+to run a SQLite database. To make them connect to Cassandra instead, you can
+either copy the Cassandra configuration into the default configuration or
+change the `bin/collector` and `bin/query` scripts to run the Cassandra
+configuration. The configuration files are located at
+`zipkin-collector-service/config' and `zipkin-query-service/config`,
+respectively. The default configurations end in `-dev.scala` and the Cassandra
+configurations end in `-cassandra.scala`.
+
+12. Run it!
 
 bin/collector
 bin/query
 bin/web
 
-12. browse to http://localhost:8080/
+browse to http://localhost:8080/
