+++ zipkin-collector-service/config/collector-cassandra.scala
@@ -14,16 +14,16 @@
  * limitations under the License.
  */
 import com.twitter.zipkin.builder.Scribe
-import com.twitter.zipkin.anormdb.{StorageBuilder, IndexBuilder}
-import com.twitter.zipkin.storage.anormdb.DB
+import com.twitter.zipkin.cassandra
 import com.twitter.zipkin.collector.builder.CollectorServiceBuilder
 import com.twitter.zipkin.storage.Store
 
-val db = DB()
-val anormBuilder = Store.Builder(
-  StorageBuilder(db),
-  IndexBuilder(db)
+val keyspaceBuilder = cassandra.Keyspace.static(nodes = Set("localhost"))
+val cassandraBuilder = Store.Builder(
+  cassandra.StorageBuilder(keyspaceBuilder),
+  cassandra.IndexBuilder(keyspaceBuilder),
+  cassandra.AggregatesBuilder(keyspaceBuilder)
 )
 
 CollectorServiceBuilder(Scribe.Interface(categories = Set("zipkin")))
-  .writeTo(anormBuilder)
+  .writeTo(cassandraBuilder)
